Make API keys rotating
API keys are limited in the number of requests they can handle per
second, it is assumed that this number is around 30 or 40.
Each account can get up to 10 API keys.
No limit of accounts is imposed.

Rotating keys allows us to send more requests per sec on
the same account.

From this commit, each time my API reaches the BS API, we
change the key.

Keys are stored in a string, seperated by a #, in a dotenv file.