Lengthen several Charfields fields limits
player / club tags :
I thought that 9 was the maximum, but found many occurrences of 10.
To give me some more room for the future, I increased the fields' length
to 15.

club description :
I thought that 100 was the maximum, but found many occurrences of more.
To give me some more room for the future, I increased the fields' length
to 200.

Match ID:
Because Player tag is longer than expected, I increases it as well