Add get_new_players command
This commandn checks all existing clubs in database and adds all players
that are not in the database yet.
However it doesn't fetch the player's data from the web, it just adds
the player to the database with the default values.

To fetch players' data, we still use update_all_players command.