Change the start tracking date from a static default to dynamic
Player base while in dev didn't change, and thus the code to compute
a player's regularity assumed the same date as a start to track
matches from.
Now, as new players will be created daily, it wouldn't be fair to
assume that they started playing on the same date as the first
player.