Reduce number of db access from create_matches_from_battlelog
Loading in memory the list of maps and brawlers in the database
is cheap enough so was can cache it in memory and avoid db access
for each battlelog match.
Potentially, this function could call 2 times the db for each
match, which is a lot of db access for nothing.