Fix UI behind the navbar
The navbar had an absolute position that would take it out
of the flow of the page. This meant that the content would
be behind the navbar. This commit fixes that by making the
navbar a block element and making the home element relative
with a negative marging to compensate.